Promote past bugs from CLAUDE.md prose into enforced lint rules
CLAUDE.md's own protocol says to prefer promoting recurring lessons into
enforced artifacts over growing a prose list — but every lesson so far
had stayed prose, and prose has never once failed a build. This applies
Merge0's hardening hierarchy (LintRule > RegressionTest >
IntentAmendment) to Merge0 itself.

crates/merge0-e2e/tests/repo_hygiene.rs adds three rules, each a bug that
actually happened:
- SQL SUM() must be cast ::BIGINT (Postgres returns NUMERIC; decoding it
  as i64 panics at runtime and it bit us twice, in telemetry and again in
  the budget/escalation queries)
- raw auth_header values keep the "Bearer " prefix, so uses must go
  through require_bearer or strip it (the broker 401'd valid runner keys)
- every adapter must emit at least one golden signal at or above the
  SHIPPED gate floor, or the source is silently dead — every signal
  guard-skipped before triage, which is how slack and asana shipped

Each rule was verified to FAIL on the reintroduced bug before being
trusted, and the three promoted lessons are deleted from the fence rather
than restated. reports.rs gains a redundant ::BIGINT so rule 1 holds with
zero exceptions to misjudge. New invariant 9 states the discipline: a
fix isn't done until the prevention is encoded, preferring a rule that
fails the build over a line nobody reads.

Also fixes a real flake the full-suite run exposed: the CLI-model test
stubs raced sibling threads' fork/exec and failed with ETXTBSY. The
helper now probes until the stub is exec-able — a deterministic wait
instead of a retry-and-hope.

+++ crates/merge0-store/src/reports.rs
@@ -329,7 +329,7 @@ impl TenantStore {
             format!(
                 "UPDATE {t} SET status = $2, dismiss_reason = $3, decided_at = $4,
                      dismissal_affected_count = (
-                         SELECT COALESCE(SUM(s.affected_count), 0)
+                         SELECT COALESCE(SUM(s.affected_count), 0)::BIGINT
                          FROM {rs} rs JOIN {signals} s ON s.id = rs.signal_id
                          WHERE rs.report_id = $1
                      )
